♻️ remove usehooks-ts and use native hook instead
