+++ developer-extension/package.json
@@ -21,7 +21,6 @@
     "@mantine/hooks": "5.0.2",
     "react": "18.0.0",
     "react-dom": "18.0.0",
-    "react-json-view": "1.21.3",
-    "usehooks-ts": "^2.9.1"
+    "react-json-view": "1.21.3"
   }
 }
