+++ developer-extension/src/panel/flushEvents.ts
@@ -1,0 +1,13 @@
+import { evalInWindow } from './evalInWindow'
+
+export function flushEvents() {
+  evalInWindow(
+    `
+      const descriptor = Object.getOwnPropertyDescriptor(Document.prototype, 'visibilityState')
+      Object.defineProperty(Document.prototype, 'visibilityState', { value: 'hidden' })
+      document.dispatchEvent(new Event('visibilitychange', { bubbles: true }))
+      Object.defineProperty(Document.prototype, 'visibilityState', descriptor)
+      document.dispatchEvent(new Event('visibilitychange', { bubbles: true }))
+    `
+  ).catch((error) => console.error('Error while flushing events:', error))
+}
